Split merge models to several files
Also removes `addDecision` from `CellMergeModel`.
