+++ nbdime-web/src/merge/model/index.ts
@@ -1,0 +1,15 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+'use strict';
+
+export {
+  NotebookMergeModel
+} from './notebook';
+
+export {
+  CellMergeModel
+} from './cell';
+
+export {
+  MetadataMergeModel
+} from './metadata';
